+++ CHANGELOG.md
@@ -16,6 +16,13 @@
   as `asdf.tagged.TaggedDict`. [[#338]](https://github.com/BAMWelDX/weldx/pull/338)
 - refactor `asdf_json_repr` into `view_tree` [[#339]](https://github.com/BAMWelDX/weldx/pull/339)
 
+### ASDF
+
+- add `uuid-1.0.0.yaml` schema as basic version 4 UUID
+  implementation [[#330]](https://github.com/BAMWelDX/weldx/pull/330)
+- add `core/graph/di_node`, `core/graph/di_edge` & `core/graph/di_graph` for implementing a
+  generic `networkx.DiGraph` [[#330]](https://github.com/BAMWelDX/weldx/pull/330)
+
 ## 0.3.3 (30.03.2021)
 
 This is a bugfix release to correctly include the asdf schema files in conda